requirements.txt
# FastAPI & Server
fastapi==0.129.0
uvicorn==0.40.0
starlette==0.52.1

# Database
pymongo==4.16.0

# Authentication & Security
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.5.0
python-dotenv==1.2.1

# Email Validation
email-validator==2.3.0

# AI/LLM Integration
google-generativeai==0.8.6

# Data Validation
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0

# Utilities
anyio==4.12.1
h11==0.16.0
idna==3.11
dnspython==2.8.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0

# Code Quality & Analysis (Optional - for development)
# radon==6.0.1
# Pygments==2.19.2

# Deprecated/Fallback dependencies (keep for compatibility)
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
pyasn1==0.6.2
rsa==4.9.1
six==1.17.0
